+++ l10n_fr_siret/models/partner.py
