--- l10n_fr_siret/partner_view.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <!-- Add the SIREN and RC -->
-        <record model="ir.ui.view" id="view_partner_add">
-            <field name="name">res.partner.form.siret</field>
-            <field name="model">res.partner</field>
-            <field name="inherit_id" ref="account.view_partner_property_form" />
-            <field name="arch" type="xml">
-                <page name="accounting" position="inside">
-                    <group name="l10n_fr_siret">
-                        <group col="4" name="siret">
-                            <!-- oe_read_only hides the field in read mode in v8+ -->
-                            <field name="siret" colspan="4" class="oe_read_only" />
-                            <field name="siren" class="oe_edit_only"
-                                string="SIREN/NIC" colspan="3"/>
-                            <field name="nic" class="oe_edit_only" nolabel="1"/>
-                        </group>
-                        <group name="rcs">
-                            <field name="company_registry" />
-                        </group>
-                    </group>
-                </page>
-            </field>
-        </record>
-    </data>
-</openerp>
