+++ l10n_fr_siret/views/company.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-    Copyright (C) 2014-2015 Akretion (http://www.akretion.com/)
-    @author: Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
+  © 2014-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 
-<openerp>
-<data>
+<odoo>
+
 
 <record id="res_company_form_l10n_fr" model="ir.ui.view">
     <field name="name">l10n_fr_siret.res.company.form</field>
@@ -20,5 +19,5 @@
     </field>
 </record>
 
-</data>
-</openerp>
+
+</odoo>
